Move logger to util module
