+++ src/main.rs
@@ -1,5 +1,4 @@
 pub mod commands;
-pub mod logger;
 pub mod project;
 pub mod template;
 pub mod util;
