+++ src/util/mod.rs
@@ -1,2 +1,3 @@
 pub mod file;
+pub mod logger;
 pub mod url;
